backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
astropy==6.0.0
pillow==10.2.0
numpy==1.26.3
scikit-image==0.22.0

